@@ -4,8 +4,8 @@ import android.content.Context
 import android.net.Uri
 import android.os.AsyncTask
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.widget.AppCompatButton
+import androidx.fragment.app.Fragment
+import androidx.appcompat.widget.AppCompatButton
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
