@@ -1,8 +1,8 @@
 package com.masterwok.simpletorrentandroid.extensions
 
 import android.content.pm.PackageManager
-import android.support.v4.app.ActivityCompat
-import android.support.v7.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.appcompat.app.AppCompatActivity
 
 
 /**
